requirements.txt
django 
djangorestframework
python-dotenv
django-cors-headers
varname 
mysqlclient
pytz
pillow
